fixed #3

+++ login/views.py
@@ -87,6 +87,7 @@ def set_password(request):
         return HttpResponse()
     return HttpResponseNotFound()
 
+
 def remove(request):
     """
     移除用户
@@ -101,7 +102,7 @@ def remove(request):
         if not request.user.is_authenticated:
             # 未登录 401
             return HttpUnauthorized()
-        if not (request.user.is_superuser and request.user.is_staff):
+        if not (request.user.is_superuser or request.user.is_staff):
             # 权限不足 403
             return HttpResponseForbidden()
         try:
@@ -118,6 +119,9 @@ def remove(request):
         if user.is_superuser:
             # 权限不足 403
             return HttpResponseForbidden()
+        if user.is_staff and not request.user.is_superuser:
+            # 权限不足 403
+            return HttpResponseForbidden()
         user.delete()
         return HttpResponse()
     return HttpResponseNotFound()
@@ -128,7 +132,7 @@ def set_staff(request):
     修改用户管理员权限
     URL: /api/set_staff/
     Method: POST
-    Permission: 管理员
+    Permission: 超级管理员
     Param:
         username: 被修改用户名
         flag: true-设置 false-取消
@@ -137,7 +141,7 @@ def set_staff(request):
         if not request.user.is_authenticated:
             # 未登录 401
             return HttpUnauthorized()
-        if not request.user.is_staff:
+        if not request.user.is_superuser:
             # 权限不足 403
             return HttpResponseForbidden()
         try:
